+++ flow/benchmarks/baselines/bottleneck2.py
@@ -32,86 +32,97 @@ DISABLE_TB = True
 DISABLE_RAMP_METER = True
 AV_FRAC = .10
 
-vehicles = Vehicles()
-vehicles.add(veh_id="human",
-             speed_mode=9,
-             routing_controller=(ContinuousRouter, {}),
-             lane_change_mode=0,
-             num_vehicles=1 * SCALING)
 
-controlled_segments = [("1", 1, False), ("2", 2, True), ("3", 2, True),
-                       ("4", 2, True), ("5", 1, False)]
-num_observed_segments = [("1", 1), ("2", 3), ("3", 3),
-                         ("4", 3), ("5", 1)]
-additional_env_params = {
-    "target_velocity": 40,
-    "disable_tb": True,
-    "disable_ramp_metering": True,
-    "controlled_segments": controlled_segments,
-    "symmetric": False,
-    "observed_segments": num_observed_segments,
-    "reset_inflow": False,
-    "lane_change_duration": 5,
-    "max_accel": 3,
-    "max_decel": 3,
-    "inflow_range": [1000, 2000]
-}
+def bottleneck2_baseline(num_runs, sumo_binary="sumo-gui"):
+    vehicles = Vehicles()
+    vehicles.add(veh_id="human",
+                 speed_mode=9,
+                 routing_controller=(ContinuousRouter, {}),
+                 lane_change_mode=0,
+                 num_vehicles=1 * SCALING)
 
-# flow rate
-flow_rate = 1900 * SCALING
+    controlled_segments = [("1", 1, False), ("2", 2, True), ("3", 2, True),
+                           ("4", 2, True), ("5", 1, False)]
+    num_observed_segments = [("1", 1), ("2", 3), ("3", 3),
+                             ("4", 3), ("5", 1)]
+    additional_env_params = {
+        "target_velocity": 40,
+        "disable_tb": True,
+        "disable_ramp_metering": True,
+        "controlled_segments": controlled_segments,
+        "symmetric": False,
+        "observed_segments": num_observed_segments,
+        "reset_inflow": False,
+        "lane_change_duration": 5,
+        "max_accel": 3,
+        "max_decel": 3,
+        "inflow_range": [1000, 2000]
+    }
 
-# percentage of flow coming out of each lane
-inflow = InFlows()
-inflow.add(veh_type="human", edge="1",
-           vehs_per_hour=flow_rate,
-           departLane="random", departSpeed=10)
+    # flow rate
+    flow_rate = 1900 * SCALING
 
-traffic_lights = TrafficLights()
-if not DISABLE_TB:
-    traffic_lights.add(node_id="2")
-if not DISABLE_RAMP_METER:
-    traffic_lights.add(node_id="3")
+    # percentage of flow coming out of each lane
+    inflow = InFlows()
+    inflow.add(veh_type="human", edge="1",
+               vehs_per_hour=flow_rate,
+               departLane="random", departSpeed=10)
 
-additional_net_params = {"scaling": SCALING}
-net_params = NetParams(in_flows=inflow,
-                       no_internal_links=False,
-                       additional_params=additional_net_params)
+    traffic_lights = TrafficLights()
+    if not DISABLE_TB:
+        traffic_lights.add(node_id="2")
+    if not DISABLE_RAMP_METER:
+        traffic_lights.add(node_id="3")
 
-sumo_params = SumoParams(
-    sim_step=0.5,
-    sumo_binary="sumo-gui",
-    print_warnings=False,
-    restart_instance=False,
-)
+    additional_net_params = {"scaling": SCALING}
+    net_params = NetParams(in_flows=inflow,
+                           no_internal_links=False,
+                           additional_params=additional_net_params)
 
-env_params = EnvParams(
-    evaluate=True,  # Set to True to evaluate traffic metrics
-    warmup_steps=40,
-    sims_per_step=1,
-    horizon=HORIZON,
-    additional_params=additional_env_params,
-)
+    sumo_params = SumoParams(
+        sim_step=0.5,
+        sumo_binary=sumo_binary,
+        print_warnings=False,
+        restart_instance=False,
+    )
 
-initial_config = InitialConfig(
-    spacing="uniform",
-    min_gap=5,
-    lanes_distribution=float("inf"),
-    edges_distribution=["2", "3", "4", "5"],
-)
+    env_params = EnvParams(
+        evaluate=True,  # Set to True to evaluate traffic metrics
+        warmup_steps=40,
+        sims_per_step=1,
+        horizon=HORIZON,
+        additional_params=additional_env_params,
+    )
 
-scenario = BottleneckScenario(name="bay_bridge_toll",
-                              generator_class=BottleneckGenerator,
-                              vehicles=vehicles,
-                              net_params=net_params,
-                              initial_config=initial_config,
-                              traffic_lights=traffic_lights)
+    initial_config = InitialConfig(
+        spacing="uniform",
+        min_gap=5,
+        lanes_distribution=float("inf"),
+        edges_distribution=["2", "3", "4", "5"],
+    )
 
-env = DesiredVelocityEnv(env_params, sumo_params, scenario)
+    scenario = BottleneckScenario(name="bay_bridge_toll",
+                                  generator_class=BottleneckGenerator,
+                                  vehicles=vehicles,
+                                  net_params=net_params,
+                                  initial_config=initial_config,
+                                  traffic_lights=traffic_lights)
 
-exp = SumoExperiment(env, scenario)
+    env = DesiredVelocityEnv(env_params, sumo_params, scenario)
 
-num_runs = 2
-results = exp.run(num_runs, HORIZON)
-avg_outflow = np.mean([outflow[-1] for outflow in results["per_step_returns"]])
-print('The average outflow over 500 seconds '
-      'across {} runs is {}'.format(num_runs, avg_outflow))
+    exp = SumoExperiment(env, scenario)
+
+    results = exp.run(num_runs, HORIZON)
+    avg_outflow = np.mean([outflow[-1]
+                           for outflow in results["per_step_returns"]])
+
+    return avg_outflow
+
+
+if __name__ == "__main__":
+    runs = 2  # number of simulations to average over
+    res = bottleneck2_baseline(num_runs=runs)
+
+    print('---------')
+    print('The average outflow over 500 seconds '
+          'across {} runs is {}'.format(runs, res))
